test: collocate dispatch interaction press coverage
